+++ app/controllers/element_judge_details_controller.rb
@@ -1,0 +1,3 @@
+class ElementJudgeDetailsController < ApplicationController
+  include ControllerConcerns::Index
+end