+++ app/decorators/element_judge_detail_decorator.rb
@@ -1,0 +1,2 @@
+class ElementJudgeDetailDecorator < EntryDecorator
+end