+++ app/decorators/score_decorator.rb
@@ -1,5 +1,4 @@
 class ScoreDecorator < EntryDecorator
-
   def ranking
     #h.link_to_score(model.ranking, model)
     h.link_to_competition(model.ranking, model.competition, category: model.category, segment: model.segment, ranking: model.ranking)