+++ app/models/official.rb
@@ -1,0 +1,4 @@
+class Official < ApplicationRecord
+  belongs_to :panel
+  belongs_to :performed_segment
+end